fix(05): change TileConfig.maxZoom from int to double for flutter_map compatibility

+++ lib/core/constants/tile_config.dart
@@ -13,5 +13,5 @@ abstract final class TileConfig {
   static const String userAgentPackageName = 'com.tagme.tagme';
 
   /// Maximum zoom level supported by Stadia Maps OSM Bright tiles.
-  static const int maxZoom = 20;
+  static const double maxZoom = 20;
 }
